Moved BlockButtonView from the view directory.
